+++ Woof/Woof/Model/KeyValueStorage/KeyValueStorage+Name.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension KeyValueStorage {
+    /// Provides the names for storage
+    enum Name {
+        /// The storage name to store information about current owner.
+        static let currentOwner = "CurrentOwnerStorage"
+    }
+}